jwt auth controller added

+++ backend/src/main/java/com/starterkit/backend/user/controller/UserController.java
@@ -2,7 +2,9 @@ package com.starterkit.backend.user.controller;
 
 import java.util.List;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -10,58 +12,76 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.starterkit.backend.security.CustomUserPrincipal;
 import com.starterkit.backend.user.dto.UserRequestByIdDto;
-import com.starterkit.backend.user.dto.UserRequestDto;
 import com.starterkit.backend.user.dto.UserResponseDto;
 import com.starterkit.backend.user.dto.UserUpdateRequestDto;
+import com.starterkit.backend.user.mapper.UserMapper;
+import com.starterkit.backend.user.service.AuthService;
 import com.starterkit.backend.user.service.UserService;
 
 import lombok.RequiredArgsConstructor;
-
 @RestController
-@RequestMapping("/api/auth")
+@RequestMapping("/api/users")
 @RequiredArgsConstructor
 public class UserController {
 
     private final UserService userService;
+    private final AuthService authService;
+    private final UserMapper userMapper;
 
-    @PostMapping("/register")
-    public ResponseEntity<UserResponseDto> register(@RequestBody UserRequestDto registerDto) {
-        UserResponseDto user = userService.register(registerDto);
-        return ResponseEntity.ok(user);
+    @GetMapping("/me")
+    public ResponseEntity<UserResponseDto> getCurrentUser(@AuthenticationPrincipal CustomUserPrincipal principal) {
+        UserResponseDto userDto = userMapper.toDto(principal.getUser());
+        return ResponseEntity.ok(userDto);
     }
 
     @GetMapping("/getAllUsers")
     public ResponseEntity<List<UserResponseDto>> getAll() {
+        if (!authService.isAdmin()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         List<UserResponseDto> users = userService.getAll();
-        
         return ResponseEntity.ok(users);
     }
 
     @PostMapping("/getById")
     public ResponseEntity<UserResponseDto> getById(@RequestBody UserRequestByIdDto userId) {
-        UserResponseDto user = userService.getById(userId);
-        if (user == null) {
-            return ResponseEntity.notFound().build();
+        if (!authService.isAdmin()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }
-        return ResponseEntity.ok(user);
+            UserResponseDto user = userService.getById(userId);
+            if (user == null) {
+                return ResponseEntity.notFound().build();
+            }
+            return ResponseEntity.ok(user);
     }
 
     @PostMapping("/update")
-    public ResponseEntity<UserResponseDto> update(@RequestBody UserUpdateRequestDto userDto) {
-        UserResponseDto updatedUser = userService.update(userDto);
-        if (updatedUser == null) {
-            return ResponseEntity.notFound().build();
-        }
-        return ResponseEntity.ok(updatedUser);
+    public ResponseEntity<UserResponseDto> update(@RequestBody UserUpdateRequestDto userDto, 
+        @AuthenticationPrincipal CustomUserPrincipal principal) {
+            // Check if user is not an admin and trying to update someone else's account
+            // if (!authService.isAdmin() && !principal.getUser().getId().equals(userDto.getId())) {
+            //     return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            // }
+
+            UserResponseDto updatedUser = userService.update(userDto);
+            if (updatedUser == null) {
+                return ResponseEntity.notFound().build();
+            }
+            return ResponseEntity.ok(updatedUser);
     }
 
     @DeleteMapping("/delete")
-    public ResponseEntity<UserResponseDto> delete(@RequestBody UserRequestByIdDto userId) {
-        UserResponseDto deletedUser = userService.delete(userId);
-        if (deletedUser == null) {
-            return ResponseEntity.notFound().build();
-        }
-        return ResponseEntity.ok(deletedUser);
+    public ResponseEntity<UserResponseDto> delete(@RequestBody UserRequestByIdDto userId, 
+        @AuthenticationPrincipal CustomUserPrincipal principal) {
+            if (!authService.isAdmin()) {
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
+            UserResponseDto deletedUser = userService.delete(userId);
+            if (deletedUser == null) {
+                return ResponseEntity.notFound().build();
+            }
+            return ResponseEntity.ok(deletedUser);
     }
 }
